get-time-python-function-app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
flask
python-dotenv
python-jose[cryptography]
flask-cors
six
azf-wsgi